+++ src/components/headers/desktop/Navigatonbar.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect, useRef, useState } from "react";
 import Link from "next/link";
-import Logo from "./Logo";
+import Logo from "../../Logo";
 import { useNavStore } from "@/stores/navigation/useNavStrore";
 import { navItems } from "@/constants/navItems";
 
@@ -42,8 +42,12 @@ export default function Navigationbar() {
   return (
     <nav className="relative">
       <div className="flex items-baseline justify-between gap-24 xl:gap-50 2xl:gap-100">
-        <Link prefetch href={navItems[0].path} onClick={() => handleClick(0, navItems[0].path)}>
-          <Logo width={84} height={36}/>
+        <Link
+          prefetch
+          href={navItems[0].path}
+          onClick={() => handleClick(0, navItems[0].path)}
+        >
+          <Logo width={84} height={36} />
         </Link>
         <div
           ref={containerRef}
